remove 1up1 down restriction

+++ RDF.Arcana.API/Features/Sales Management/Sales Transactions/AddTransaction.cs
@@ -68,8 +68,8 @@ public class AddTransaction : ControllerBase
         public class Item
         {
             public int ItemId { get; set; }
-            public int Quantity { get; set; }
-            public int UnitPrice { get; set; }
+            public decimal Quantity { get; set; }
+            public decimal UnitPrice { get; set; }
         }
     }
 
@@ -103,8 +103,8 @@ public class AddTransaction : ControllerBase
             public string ItemCode { get; set; }
             public string ItemDescription { get; set; }
             public string Uom { get; set; }
-            public int Quantity { get; set; }
-            public int UnitPrice { get; set; }
+            public decimal Quantity { get; set; }
+            public decimal UnitPrice { get; set; }
             public decimal Amount { get; set; }
         }
